Move get_two_hands() to combat_card module
